+++ src/simple_range/__init__.py
@@ -1,3 +1,3 @@
 from ._core import ALL, FIRST, SECOND, THIRD, LAST_2, LAST_1, LAST
-from ._range import Range
-from ._index import Index
+from ._range import Range, range_indices
+from ._index import Index, index_value