+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - added "simple_range.Index" class for a single index
+- added convenience methods "range_indices" and "index_value"
 
 
 0.0.1 (2023-02-21)