+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.0.3 (2023-10-19)
+------------------
+
+- added `PLACEHOLDERS` list
+- added class method `help` to `Index` and `Range` class for returning a short help string
+
+
 0.0.2 (2023-02-21)
 ------------------